# openapi .json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Define an API",
    "title": "Restful API(connect two databases)",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "url": "http://www.example.com/support",
      "email": "a167829435@gmail.com"
    }
  },
  "paths": {
...
    "/v1/mysql/getall": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "從peoples取得所有資料",
        "responses": {
          "200": {
            "description": "Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.People"
                },
                "example": {
                  "Key1": "1",
                  "Key2": "2",
                  "Key3": "3",
                  "Number": 100,
                  "Gender": 1,
                  "Birth": 83,
                  "Injury_degree": "3",
                  "Injury_position": 10,
                  "Protection": 1,
                  "Phone": 2,
                  "Person": "B03",
                  "Car": "7",
                  "Action_status": 1,
                  "Qualification": 2,
                  "License": 3,
                  "Drinking": 1,
                  "Hit": 2
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "Serve(database) error!"
              }
            }
          }
        },
        "summary": "get all data from peoples",
        "tags": [
          "People"
        ]
      }
    },
    "/v1/mysql/getsome/{key1}/{key2}/{key3}": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "從peoples取得部分資料",
        "parameters": [
          {
            "name": "key1",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key2",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key3",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.People"
                },
                "example": {
                  "Key1": "1",
                  "Key2": "2",
                  "Key3": "3",
                  "Number": 100,
                  "Gender": 1,
                  "Birth": 83,
                  "Injury_degree": "3",
                  "Injury_position": 10,
                  "Protection": 1,
                  "Phone": 2,
                  "Person": "B03",
                  "Car": "7",
                  "Action_status": 1,
                  "Qualification": 2,
                  "License": 3,
                  "Drinking": 1,
                  "Hit": 2
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "The user does not exist!"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "Serve(database) error!"
              }
            }
          }
        },
        "summary": "get some data from events",
        "tags": [
          "People"
        ]
      }
    },
    "/v1/mysql/addvalue": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "插入數值至peoples",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/models.People"
              }
            }
          },
          "description": "add data",
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.People"
                },
                "example": {
                  "Key1": "1",
                  "Key2": "2",
                  "Key3": "3",
                  "Number": 100,
                  "Gender": 1,
                  "Birth": 83,
                  "Injury_degree": "3",
                  "Injury_position": 10,
                  "Protection": 1,
                  "Phone": 2,
                  "Person": "B03",
                  "Car": "7",
                  "Action_status": 1,
                  "Qualification": 2,
                  "License": 3,
                  "Drinking": 1,
                  "Hit": 2
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "Serve(database) error!"
              }
            }
          }
        },
        "summary": "add value to peoples",
        "tags": [
          "People"
        ]
      }
    },
    "/v1/mysql/update": {
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "更新peoples數值",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/model.mysqlupdate"
              },
              "example": {
                "key1": "12345",
                "death": "0"
              }
            }
          },
          "description": "update data",
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                },
                "example": "Successfully update"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "Serve(database) error!"
              }
            }
          }
        },
        "summary": "update value",
        "tags": [
          "People"
        ]
      }
    },
    "/v1/mysql/delete/{key1}/{key2}/{key3}": {
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "刪除peoples數值",
        "parameters": [
          {
            "name": "key1",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key2",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key3",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                },
                "example": "Successfully delete"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "Serve(database) error!"
              }
            }
          }
        },
        "summary": "delete value",
        "tags": [
          "People"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:8080/v1"
    }
  ],
  "components": {
...
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "model.mysqlupdate": {
        "properties": {
          "birth": {
            "type": "integer"
          },
          "key1": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "models.Error": {
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "models.People": {
        "properties": {
          "action_status": {
            "type": "integer"
          },
          "birth": {
            "type": "integer"
          },
          "car": {
            "type": "string"
          },
          "drinking": {
            "type": "integer"
          },
          "gender": {
            "type": "integer"
          },
          "hit": {
            "type": "integer"
          },
          "injury_degree": {
            "type": "string"
          },
          "injury_position": {
            "type": "integer"
          },
          "key1": {
            "type": "string"
          },
          "key2": {
            "type": "string"
          },
          "key3": {
            "type": "string"
          },
          "license": {
            "type": "integer"
          },
          "number": {
            "type": "integer"
          },
          "person": {
            "type": "string"
          },
          "phone": {
            "type": "integer"
          },
          "protection": {
            "type": "integer"
          },
          "qualification": {
            "type": "integer"
          }
        },
        "type": "object"
      },
...
    }
  }
}